rename root -> views
