+++ packages/next/client/views-next.js
@@ -1,4 +1,4 @@
-import { hydrate, version } from './root-index'
+import { hydrate, version } from './views-index'
 
 window.next = {
   version,
